+++ Cargo.toml
@@ -12,6 +12,9 @@ ci = []
 [dependencies]
 crypto-hash = "0.3.4"
 async-trait = "0.1.68"
+rexie = "0.4"
+send_wrapper = { version = "0.6.0", features = ["futures"] }
+serde-wasm-bindgen = "0.6.5"
 serde = { version = "1.0.151", features = ["derive"] }
 serde_json = "1.0.79"
 redis = "0.24.0"
